requirements.txt
asgiref==3.2.3
atomicwrites==1.3.0
beautifulsoup4==4.6.0
bleach==3.1.1
boto3==1.11.10
botocore==1.14.10
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.10
django-appconf==1.0.3
django-avatar==5.0.0
django-classy-tags==1.0.0
django-cogwheels==0.3
django-compressor==2.4
django-contrib-comments==1.9.2
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-formtools==2.2
django-imagekit==4.0.1
django-modelcluster==5.0.1
django-sekizai==1.1.0
django-social-share==1.3.2
django-storages==1.8
django-taggit==1.2.0
django-treebeard==4.3.1
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.5
djangocms-admin-style==1.2.8
djangorestframework==3.11.0
docutils==0.15.2
draftjs-exporter==2.1.7
filebrowser-safe==0.5.0
future==0.18.2
grappelli-safe==0.5.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jmespath==0.9.4
l18n==2018.5
Mezzanine==4.3.1
oauthlib==3.1.0
Pillow==6.2.2
psycopg2-binary==2.8.4
pycparser==2.19
PyJWT==1.7.1
Pyphen==0.9.5
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
rcssmin==1.0.6
requests==2.22.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
s3transfer==0.3.2
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tinycss2==0.6.1
tzlocal==2.0.0
Unidecode==1.1.1
urllib3==1.25.8
wagtail==2.8
wagtail-gallery==0.1.1
wagtailmenus==3.0
WeasyPrint==45
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.3
pandas
django-import-export


